Add Expirmental NTRO driver
* NTRO DLDI driver is for mounting a imbedded FAT image on main rom and
is accessed via direct card reads. Note this only works on certain
flashcarts like N-Card and homebrew flashcarts like the upcoming stm32
cart. The fat image can be added to NitroFS of rom but can be manually
appended to a rom provided that used rom size in header is updated to
account for it. (easier to just use ndstool to repack it into nitrofs
though).
* compile flags added to gbcf to allow compiling driver to target a
specific slot-2 flashcard. This skips the autodetection phase. Use a
compile flag for specific cart if you have issues getting dldi init to
succeed.